fix(panel): Keep History filters while briefs load
